% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clplot.R
\name{clplot}
\alias{clplot}
\title{Make a Command-Line Plot}
\usage{
clplot(x, y, plot.size = 10)
}
\arguments{
\item{x}{A numeric vector. NAs are removed.}

\item{y}{A numeric vector. NAs are removed.}

\item{plot.size}{Integer; number of bins.}
}
\description{
\code{clplot} Makes a plot in the command line. Observations are split linearly into three groups ., x & X by frequency.
}
\examples{
clplot(1:100, 1:100)


##     1 12 23 34 45 56 67 78 89 100
## 100                           x
## 89                         x
## 78                      x
## 67                   x
## 56                x
## 45             x
## 34          x
## 23       x
## 12    x
## 1   x

}
