% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcols.R
\name{ggcols}
\alias{ggcols}
\title{Generate the ggplot2 color wheel}
\usage{
ggcols(n)
}
\arguments{
\item{n}{Integer; the number of colours to get.}
}
\description{
\code{ggcols} Returns a vector of ggplot2 colours. Inspired by an answer by Hadley, Colby J. & zx8754 on Cross Validated.
}
\examples{
ggcols(1)
## "#F8766D"
}
