% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.evenodd.R
\name{is.odd}
\alias{is.odd}
\title{Check if a vector of numbers is Odd}
\usage{
is.odd(x, n = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer specifying if a number of odd numbers should be returned.}
}
\description{
\code{is.odd} Returns a vector of TRUE/FALSEs by applying x%2 to a vector x.
}
\examples{
is.odd(1:5)

## TRUE FALSE  TRUE FALSE TRUE
}
