% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_engine.R
\name{ParseTemplate}
\alias{ParseTemplate}
\title{ParseTemplate}
\usage{
ParseTemplate()
}
\value{
This function outputs an example input to the SSMet routine.
}
\description{
ParseTemplate
}
\examples{
SSMet(ParseTemplate())
}
