% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_fancy.R
\name{qqplot_fancy}
\alias{qqplot_fancy}
\title{Make a pretty QQ-Plot}
\usage{
qqplot_fancy(sample, dist = "normal", col = sample(ggcols(50)[c(1:5, 17:38,
  47:49)], 1), start = NULL, xlab = NULL, ylab = NULL, title = NULL)
}
\arguments{
\item{sample}{A numeric vector with the "actual" observations.}

\item{dist}{Either (1) a string (default "normal") specifying a known distribution Or (2) another vector of theoretical/model observations. The distribution can be one of beta, exponential, gamma, log-normal, lognormal, normal, pareto, t or weibull.}

\item{col}{(optional) A string specifying point colour. Defaults to a random colour from the ggplots color wheel.}

\item{start}{(optional) A list of named starting/initial values for fitditr.}

\item{xlab}{(optional) A string containing the x axis label.}

\item{ylab}{(optional) A string containing the y axis label.}

\item{title}{(optional) A string containing the title.}
}
\description{
\code{qqplot_fancy} Extends the qqplot() function by using ggplot2 for making a pretty qqplot and adding a best-fitting distribution functionality.
}
\examples{
qqplot_fancy(rnorm(1000)) ## pretty line
qqplot_fancy(rexp(1000), "exponential") ## pretty line
qqplot_fancy(rexp(1000), rexp(1000)) ## pretty line
qqplot_fancy(rnorm(10), col = "blue") ## pretty blue line
}
