% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_engine.R
\name{SSMet}
\alias{SSMet}
\title{Simulate an ECDF using Subset Simulation}
\usage{
SSMet(parse_string, data = NULL)
}
\arguments{
\item{parse_string}{A string containing the simulation schema. For an example, use \code{ParseTemplate()}}

\item{data}{(optional) A named list of objects that's passed onto the SSMet function.}
}
\value{
A data frame with the output function, the corresponding simulated inputs and stage descriptions ("NMC" or "MetMCMC_ChainX_StepX").
}
\description{
Simulate an ECDF using Subset Simulation
}
\examples{
SSMet(ParseTemplate())
}
